bot/requirements.txt
fastapi==0.100.0
passlib==1.7.4
pydantic==2.0.3
rich==13.5.2
pydub==0.25.1
python-dotenv==0.21.0
PyYAML==6.0
requests==2.31.0
pymongo==3.12.3
tiktoken==0.4.0
uvicorn==0.23.1
gunicorn==21.2.0
starlette==0.27.0
Babel==2.12.1
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
python-telegram-bot[rate-limiter]==20.1
pytz==2023.3



